Unset and reset environment variables before/after
These tests were always passing because the relevant environment variables
were always missing on my machine locally. This was because the `ENV` was not
populated.

If I ran the tests with `foreman run`, the tests failed.

Ideally, we should remember to run tests with `foreman run bundle exec
rspec` so that the ENV vars are populated.