Make sure Mailchimp requests are matched with VCR
The suffix `-us10` tells Mailchimp which server use. When we record VCR
cassettes they are recorded against that server so we must use the same
server here.

This is bad.

I tried to make the request matching more lenient, but could not figure out how
to configure Cucumber and VCR to make it happen.