Don't add members via Eventbrite
Adding members to Capsule via Eventbrite ticket purchases was causing
more problems than it was solving.

People would signup using their personal name in Eventbrite and this
value would propagated to Capsule and then the member directory.

However, if you wanted to change the person's name to a company name it
would work for a while until the Eventbrite job ran. When that job ran
it would try to add a new member to Capsule because people are fetched from
Capsule using their company name (for organizations). This would fail
because of the database constraint on email.

We're disabling this functionality as the amount of tickets that come
through via Eventbrite is minimal so the team will manually create the
entry in Capsule as necessary.