Add new processors for invoicing from Eventbrite
This can’t be used at the moment as we have no way of knowing if an
order is already paid or needs an invoice.

We may need to hook into paypal/xero apis somehow.

Instead of now checking for events the processor requests all new orders
since the beginning of the day and generates the information for the
`Invoicer` from an order’s event and attendees.