Stop the booking value counting renewal months
This was causing the value to be higher than it should be.

Each transaction has either 'Signup' or 'Renewal' in it's memo field.
The first payment of a monthly subscription always has 'Signup' in the
memo field. Subsequent payments have 'Renewal'.

We only include 'Signup' transactions for the booking value.