Big refactor of the booking value report
We had to move to using subscription data rather than transaction data
to generate the report the transaction data could not be used to count
free products as no transaction would exist.

As a side effect of using subscriptions, the report is now much simplier
and should be more robust.